Move `sign_zone()` to `signer/full.rs`
